Paginação e agendamento de monitor corrigido!

+++ monitor/views.py
@@ -1,12 +1,11 @@
 from django.shortcuts import render, redirect, get_object_or_404
 from django.db.models import Q
 from django.contrib import messages
-from datetime import datetime
-from django.db import transaction
 from .forms import AgendamentoMonitorForm
 from secretaria.models import Agendamento, Tipo
 from laboratorio.models import Laboratorio
 from django.contrib.auth.decorators import login_required, user_passes_test
+from django.core.paginator import Paginator
 
 def is_monitor(user):
     return user.groups.filter(name='Monitor').exists()
@@ -38,50 +37,56 @@ def agendar_monitor(request):
                         (Q(hora_inicio__lte=f.hora_inicio) & Q(hora_fim__gte=f.hora_inicio)) |
                         (Q(hora_inicio__lte=f.hora_fim) & Q(hora_fim__gte=f.hora_fim))
                 )
-            )
+            ).exclude(status='Cancelado')
 
             if coliding_events.exists():
-                if f.tipo.titulo == 'Aula':
-                    messages.error(request, 'Já existe um agendamento conflitante neste horário!')
-                elif f.tipo.titulo == 'Monitoria' and coliding_events.filter(tipo__titulo='Aula').exists():
-                    messages.error(request, 'Já existe uma aula cadastrada nesse horário')
-                else:
-                    # Se for outra monitoria ou não houver conflito, salva o novo agendamento
+                if coliding_events.filter(tipo__titulo='Monitoria').exists():
                     f.user = request.user
                     f.status = 'Solicitado'
+                    f.titulo = f.titulo.capitalize()
                     f.tipo = get_object_or_404(Tipo, titulo='Monitoria')
                     f.save()
                     messages.success(request, 'Solicitação de agendamento realizado com sucesso!')
-
+                    return redirect('monitor:lista_monitor')
+                else:
+                    messages.error(request, 'Já existe uma agendamento nesse horário!')
             else:
                 f.user = request.user
                 f.status = 'Solicitado'
+                f.titulo = f.titulo.capitalize()
                 f.tipo = get_object_or_404(Tipo, titulo='Monitoria')
                 f.save()
                 messages.success(request, 'Solicitação de agendamento realizado com sucesso!')
                 return redirect('monitor:lista_monitor')
-        else:
-            context['form'] = form
-
     else:
         form = AgendamentoMonitorForm()
-        context['form'] = form
+    context['form'] = form
     return render(request, template_name, context)
-
 
 """@login_required(login_url='/contas/login/')
 @user_passes_test(is_monitor)"""
 def listar_monitor(request):
-    agendamentos = Agendamento.objects.all()   #filter(user=request.user)
-    return render(request, 'monitor/lista_agendamento_monitor.html', {'agendamentos': agendamentos})
+    template_name = 'monitor/lista_agendamento_monitor.html'
+    consulta = Agendamento.objects.all()   #filter(user=request.user)
+    paginator = Paginator(consulta, 3)
+
+    page_number = request.GET.get("page")
+    agendamentos = paginator.get_page(page_number)
+    context = {
+        'agendamentos': agendamentos
+    }
+    return render(request, template_name, context)
 
 """@login_required(login_url='/contas/login/')
 @user_passes_test(is_monitor)"""
 def excluir_agendamento(request, agendamento_id):
     try:
         agendamento = Agendamento.objects.get(id=agendamento_id)
-        agendamento.delete()
-        messages.success(request, "Agendamento excluído!")
+        if agendamento.status != 'Agendado':
+            agendamento.delete()
+            messages.success(request, "Agendamento excluído!")
+        else:
+            messages.error(request, 'Essa solicitação estar agendada por esse motivo não pode ser excluída!')
     except Agendamento.DoesNotExist:
         messages.error(request, "Agendamento não encontrado.")
